re-structured functions and files for getSpecificPost
