+++ src/js/index.mjs
@@ -2,11 +2,12 @@
 
 import { registerUser } from "./auth/registration.mjs";
 import { loginUser } from "./auth/login.mjs";
-import { displaySpecificPost } from "./posts/getSpecificPost.mjs";
+//import { displaySpecificPost } from "./posts/getSpecificPost.mjs";
 import { setup } from "./posts/setup.mjs";
 import { filterTags } from "./filter/filterTags.mjs";
 import { sortPosts } from "./sorter/sortType.mjs";
 import { signout } from "./auth/logout.mjs";
+import { setupSpecificPost } from "./posts/setupSpecific.mjs";
 
 const path = window.location.pathname;
 console.log(path);
@@ -19,7 +20,8 @@ if (path === "/index.html") {
   setup();
   signout();
 } else if (path === "/post-specific.html") {
-  displaySpecificPost();
+  //displaySpecificPost();
+  setupSpecificPost();
   signout();
 } else if (path === "/post-filter.html") {
   filterTags();
